+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
 * Added `securepay` to the payment processor validation.
 * Added `credit_application` and `fund_transfer` to event type validation.
+* Added the input `/event/party`. This is the party submitting the
+  transaction.
 
 3.3.0 (2025-05-23)
 ------------------